// ir-gui/src/app/components/settings/mvno-management/mvno-management.interface.ts
export interface IMvnoManagement {
  description: string;
  email: string;
  id: number;
  fullName: string;
  lastModifiedById: number;
  lastModifiedByName: string;
  mvnoPaymentDueDays: number;
  name: string;
  password: string;
  passwordPolicyId: number;
  phone: string;
  profileImage: string;
  roleId: number;
  status: string;
  username: string;
}